Moving `ThemeProvider` into the `VideoPlayer`
