+++ src/components/video-player/VideoPlayer.tsx
@@ -1,7 +1,14 @@
-import { Theme } from '@mui/material/styles';
+import CssBaseline from '@mui/material/CssBaseline';
+import {
+	StyledEngineProvider,
+	Theme,
+	ThemeProvider,
+} from '@mui/material/styles';
+import { deepmerge } from '@mui/utils';
 import { FC } from 'react';
 
 import { VideoProvider } from '../../context/video';
+import { createPlayerTheme } from '../../theme';
 import { ControlsConfig } from '../../types';
 import { DEFAULT_CONTROLS_CONFIG } from '../controls/controls-config';
 import { FileActionPanelProps } from '../file-action-panel/FileActionPanel';
@@ -45,23 +52,28 @@ export const VideoPlayer: FC<VideoPlayerProps> = ({
 	const hasPlayEnabled = videoUrl === currentPlayingUrl;
 
 	const onPlay = () => setCurrentPlayingUrl?.(videoUrl);
+	const nestedThemes = deepmerge(createPlayerTheme(), theme || {});
 
 	return (
-		<VideoProvider controlsConfig={controlsConfig}>
-			<VideoContainer
-				className={className}
-				videoUrl={videoUrl}
-				hasPlayEnabled={hasPlayEnabled}
-				onPlay={onPlay}
-				onDelete={onDelete}
-				onDownload={onDownload}
-				removeAsCover={removeAsCover}
-				setAsCover={setAsCover}
-				actionPanelClassName={actionPanelClassName}
-				hasImageCover={hasImageCover}
-				isCover={isCover}
-				theme={theme}
-			/>
-		</VideoProvider>
+		<ThemeProvider theme={nestedThemes}>
+			<StyledEngineProvider injectFirst>
+				<CssBaseline />
+				<VideoProvider controlsConfig={controlsConfig}>
+					<VideoContainer
+						className={className}
+						videoUrl={videoUrl}
+						hasPlayEnabled={hasPlayEnabled}
+						onPlay={onPlay}
+						onDelete={onDelete}
+						onDownload={onDownload}
+						removeAsCover={removeAsCover}
+						setAsCover={setAsCover}
+						actionPanelClassName={actionPanelClassName}
+						hasImageCover={hasImageCover}
+						isCover={isCover}
+					/>
+				</VideoProvider>
+			</StyledEngineProvider>
+		</ThemeProvider>
 	);
 };
